requirements.txt
Appium-Python-Client==0.41
selenium==3.141.0
pyyaml==3.13